Move specific ScoringContext classes directly next to the LabelScorers that use them
