+++ src/Nn/LabelScorer/FixedContextOnnxLabelScorer.hh
@@ -23,6 +23,23 @@
 namespace Nn {
 
 /*
+ * Scoring context that describes a sequence of previously observed labels as well as the current decoding step
+ */
+struct SeqStepScoringContext : public ScoringContext {
+    std::vector<LabelIndex> labelSeq;
+    Speech::TimeframeIndex  currentStep;
+
+    SeqStepScoringContext();
+    SeqStepScoringContext(std::vector<LabelIndex> const& seq, Speech::TimeframeIndex step);
+    SeqStepScoringContext(std::vector<LabelIndex>&& seq, Speech::TimeframeIndex step);
+
+    bool   isEqual(ScoringContextRef const& other) const override;
+    size_t hash() const override;
+};
+
+typedef Core::Ref<SeqStepScoringContext const> SeqStepScoringContextRef;
+
+/*
  * Label Scorer that performs scoring by forwarding the input feature at the current timestep together
  * with a fixed-size sequence of history tokens through an ONNX model.
  * A common use case would be a neural transducer model with a fixed-size history.
