+++ src/Nn/LabelScorer/ScoringContext.hh
@@ -16,17 +16,16 @@
 #ifndef SCORING_CONTEXT_HH
 #define SCORING_CONTEXT_HH
 
-#include <optional>
-
 #include <Core/ReferenceCounting.hh>
 #include <Mm/Types.hh>
 #include <Nn/AbstractStateManager.hh>
-#include <Onnx/OnnxStateVariable.hh>
-#include <Onnx/Value.hh>
 
 #include "Types.hh"
 
 namespace Nn {
+
+size_t labelSeqHash(std::vector<LabelIndex> const& labelSeq);
+bool   labelSeqEqual(std::vector<LabelIndex> const& lhs, std::vector<LabelIndex> const& rhs);
 
 /*
  * Empty scoring context base class
@@ -53,24 +52,6 @@ struct ScoringContextEq {
 };
 
 /*
- * Combines multiple scoring contexts at once
- */
-struct CombineScoringContext : public ScoringContext {
-    std::vector<ScoringContextRef> scoringContexts;
-
-    CombineScoringContext()
-            : scoringContexts() {}
-
-    CombineScoringContext(std::vector<ScoringContextRef>&& scoringContexts)
-            : scoringContexts(scoringContexts) {}
-
-    bool   isEqual(ScoringContextRef const& other) const;
-    size_t hash() const;
-};
-
-typedef Core::Ref<CombineScoringContext const> CombineScoringContextRef;
-
-/*
  * Scoring context that only describes the current decoding step
  */
 struct StepScoringContext : public ScoringContext {
@@ -87,126 +68,6 @@ struct StepScoringContext : public ScoringContext {
 };
 
 typedef Core::Ref<StepScoringContext const> StepScoringContextRef;
-
-/*
- * Scoring context that describes a sequence of previously observed labels as well as the current decoding step
- */
-struct SeqStepScoringContext : public ScoringContext {
-    std::vector<LabelIndex> labelSeq;
-    Speech::TimeframeIndex  currentStep;
-
-    SeqStepScoringContext()
-            : labelSeq(), currentStep(0ul) {}
-    SeqStepScoringContext(std::vector<LabelIndex> const& seq, Speech::TimeframeIndex step)
-            : labelSeq(seq), currentStep(step) {}
-    SeqStepScoringContext(std::vector<LabelIndex>&& seq, Speech::TimeframeIndex step)
-            : labelSeq(std::move(seq)), currentStep(step) {}
-
-    bool   isEqual(ScoringContextRef const& other) const;
-    size_t hash() const;
-};
-
-typedef Core::Ref<SeqStepScoringContext const> SeqStepScoringContextRef;
-
-/*
- * Hidden state represented by a dictionary of named ONNX values
- */
-struct OnnxHiddenState : public Core::ReferenceCounted {
-    std::unordered_map<std::string, Onnx::Value> stateValueMap;
-
-    OnnxHiddenState()
-            : stateValueMap() {}
-
-    OnnxHiddenState(std::vector<std::string>&& names, std::vector<Onnx::Value>&& values) {
-        verify(names.size() == values.size());
-        stateValueMap.reserve(names.size());
-        for (size_t i = 0ul; i < names.size(); ++i) {
-            stateValueMap.emplace(std::move(names[i]), std::move(values[i]));
-        }
-    }
-};
-
-typedef Core::Ref<OnnxHiddenState const> OnnxHiddenStateRef;
-
-/*
- * Scoring context consisting of a hidden state.
- * Assumes that two hidden states are equal if and only if they were created
- * from the same label history.
- */
-struct OnnxHiddenStateScoringContext : public ScoringContext {
-    std::vector<LabelIndex>    labelSeq;  // Used for hashing
-    mutable OnnxHiddenStateRef hiddenState;
-    mutable bool               requiresFinalize;
-
-    OnnxHiddenStateScoringContext()
-            : labelSeq(), hiddenState(), requiresFinalize(false) {}
-
-    OnnxHiddenStateScoringContext(std::vector<LabelIndex> const& labelSeq, OnnxHiddenStateRef state, bool requiresFinalize)
-            : labelSeq(labelSeq), hiddenState(state), requiresFinalize(requiresFinalize) {}
-
-    bool   isEqual(ScoringContextRef const& other) const;
-    size_t hash() const;
-};
-
-typedef Core::Ref<OnnxHiddenStateScoringContext const> OnnxHiddenStateScoringContextRef;
-
-/*
- * Scoring context for computation of CTC prefix scores.
- * Contains the label sequence, a parent pointer for lazy computation of the prefix score,
- * time-wise prefix scores and a cache for the score of extended prefixes.
- * Hash and equality operators are based on the label sequence.
- */
-struct CtcPrefixScoringContext : public ScoringContext {
-    struct PrefixScore {
-        Score blankEndingScore    = std::numeric_limits<Score>::infinity();
-        Score nonBlankEndingScore = std::numeric_limits<Score>::infinity();
-
-        Score totalScore() const {
-            return Math::scoreSum(blankEndingScore, nonBlankEndingScore);
-        }
-    };
-
-    std::vector<LabelIndex>                           labelSeq;
-    ScoringContextRef                                 parent;            // Parent prefix without the last label, used to finalize lazily
-    mutable std::shared_ptr<std::vector<PrefixScore>> timePrefixScores;  // Represents neg-log-probabilities of emitting `labelSeq` ending in blank or nonblank after each real CTC frame t = 0, ..., T - 1
-    mutable std::optional<Score>                      prefixScore;       // Cached score of the prefix -log P(prefix), computed on demand from the parent; needed to compute score-delta for next token
-    mutable std::unordered_map<LabelIndex, Score>     extScores;         // Cache for -log P(prefix + token, ...) to avoid repeated computation
-    mutable bool                                      requiresFinalize;  // Check whether the mutable members have been set
-
-    CtcPrefixScoringContext()
-            : labelSeq(), parent(), timePrefixScores(), prefixScore(), extScores(), requiresFinalize(true) {}
-
-    CtcPrefixScoringContext(std::vector<LabelIndex>&& seq, ScoringContextRef const& parent)
-            : labelSeq(std::move(seq)), parent(parent), timePrefixScores(), prefixScore(), extScores(), requiresFinalize(true) {}
-
-    bool   isEqual(ScoringContextRef const& other) const;
-    size_t hash() const;
-};
-
-typedef Core::Ref<const CtcPrefixScoringContext> CtcPrefixScoringContextRef;
-
-/*
- * Scoring context for ONNX models with state managed by a StateManager.
- * The state stores only the slice produced for the most recent token.
- */
-struct StateManagedOnnxScoringContext : public ScoringContext {
-    using StateManager = AbstractStateManager<Onnx::Value, Onnx::OnnxStateVariable>;
-    using HistoryState = StateManager::HistoryState;
-
-    std::vector<LabelIndex>                                 labelSeq;
-    mutable Core::Ref<StateManagedOnnxScoringContext const> parent;
-
-    mutable std::shared_ptr<HistoryState> state;
-
-    StateManagedOnnxScoringContext(HistoryState&& initialState);
-    StateManagedOnnxScoringContext(std::vector<LabelIndex>&&                       labelSeq,
-                                   Core::Ref<StateManagedOnnxScoringContext const> parent);
-
-    bool   isEqual(ScoringContextRef const& other) const override;
-    size_t hash() const override;
-};
-
-typedef Core::Ref<StateManagedOnnxScoringContext const> StateManagedOnnxScoringContextRef;
 
 }  // namespace Nn
